fix(431-2): active network icon has too much margin and adding optional prop (#12405)
## **Description**

The active network icon in the header has too much margin and is
misaligned with the network title.

I have also added an optional prop of `networkName` that allows
developers to pass in the network name of their choice. This is in
preparation of
[multichain](https://github.com/MetaMask/metamask-mobile/pull/12256)
support, currently we are showing the selected network name, in
multichain we want to be able to pass in a network name and not the
selected chain.

## **Related issues**
Fixes: 

## **Manual testing steps**
1. Goto asset details, swaps, or buy. Basically any screen that has the
header with network name
2.
3.

## **Screenshots/Recordings**
(green icon at the very top)
| Before  | After  |
|:---:|:---:|

|![asset_details_ios_before](https://github.com/user-attachments/assets/26de61bb-a217-456f-9d40-8bc552f4b1ec)|![asset_details_ios_after](https://github.com/user-attachments/assets/a4d4667b-4b1f-485c-a9f9-ef78872da772)|

### **Before**

NA

### **After**

NA

## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.