test: Add ramps URL scheme deeplinking e2e (#12747)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

The purpose of this PR is to add e2e coverage for the URL schema
deeplinks in the ramps flow.

The goal here is to ensure that we properly test the URL scheme deep
linking in these flows.

### Sell Deep link

```
Scenario 1
Given I open a sell ETH deeplink for mainnet 
Then the app should launch with the ramps build quotes page
And the token I want to sell is displayed correctly
And the amount i want to send is displayed correctly

Scenario 2
Given I open a sell deeplink on an unsupported network
Then the app should launch with the ramps build quotes page
And I am prompted to add the network
When I add the network
Then the Quotes page should be displayed 
```

### Buy deeplink Flow

```
Scenario 3
Given i deeplink to the buy eth flow on mainnet
Then the app should launch with the ramps build quotes page
And the token I want to sell is displayed correctly
And the amount i want to send is displayed correctly

Scenario 4
Given i deeplink to the buy eth flow on a popular network
Then the app should launch with the ramps build quotes page
And the token I want to sell is displayed correctly
And the amount i want to send is displayed correctly

```
## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.